refactor(credentials): address mentor review — use Credo APIs, SD-JWT naming, remove id from schema
- Remove did-utils.ts and did-utils.test.ts: Credo provides DidDocument,
  VerificationMethod, resolveDidDocument (via agent.dids), and
  dereferenceKey natively — no need for custom re-implementations

- Move ajv from dependencies to devDependencies (^8.12.0): only used in
  tests, version aligns with @animo-id/pex requirements

- Remove id property from credential schema: in SD-JWT VC, holder
  identity is bound via the cnf claim (proof of possession), not a
  schema-level id field

- Rename credential-subject.spec.ts to credential-claims.spec.ts: use
  SD-JWT VC terminology (claims/payload) instead of W3C VC terminology
  (credential subject)

- Rename mock-issuer-payload.json to mock-sd-jwt-payload.json

- Update ADR-0001 to reference Credo's dereferenceKey instead of the
  removed getDeterministicSigningKey helper
