+++ heka-identity-service/package.json
@@ -71,7 +71,6 @@
     "@openwallet-foundation/askar-nodejs": "0.6.0",
     "@openwallet-foundation/askar-shared": "0.6.0",
     "@types/multer": "^1.4.11",
-    "ajv": "^8.20.0",
     "async-mutex": "^0.5.0",
     "axios": "^1.17.0",
     "chalk": "^4.1.2",
@@ -116,6 +115,7 @@
     "@types/uuid": "^9.0.1",
     "@types/ws": "^8.18.1",
     "@vitest/coverage-v8": "^3.2.6",
+    "ajv": "^8.12.0",
     "eslint": "^10.4.1",
     "eslint-config-prettier": "^10.1.8",
     "eslint-import-resolver-typescript": "^4.4.5",
