+++ heka-identity-service/src/credentials/fixtures/github-contributor-credential.schema.json
@@ -1,12 +1,8 @@
 {
   "$schema": "https://json-schema.org/draft/2020-12/schema",
-  "title": "GithubContributorCredentialSubject",
+  "title": "GithubContributorCredentialClaims",
   "type": "object",
   "properties": {
-    "id": {
-      "type": "string",
-      "pattern": "^did:hedera:.*"
-    },
     "githubUsername": {
       "type": "string",
       "minLength": 1
@@ -19,7 +15,7 @@
       "type": "string",
       "pattern": "^[0-9A-F]{16,40}$"
     }
-  }, 
-  "required": ["id", "githubUsername", "githubAccountId", "gpgFingerprint"],
+  },
+  "required": ["githubUsername", "githubAccountId", "gpgFingerprint"],
   "additionalProperties": false
 }
